refactor(vi): move vi-mode state + enter/exit actions to action/vi/mode
ui/vi_mode.rs held zero UI: it is the vi feature's enter/exit command pair
(EntityEvents + apply observers) and the ViModeState marker — the action
layer's pattern verbatim. ViModePlugin now registers under ViActionPlugin.
