+++ src/main.rs
@@ -31,7 +31,7 @@ use orzma_webview::{OrzmaWebviewPlugin, cef_plugin};
 use orzma_webview_host::WebviewAssetRegistry;
 use session::default::DefaultSessionPlugin;
 use ui::ime_overlay::ImeOverlayPlugin;
-use ui::{OrzmaUiPlugin, vi_mode::ViModePlugin, vi_mode_indicator::ViModeIndicatorPlugin};
+use ui::{OrzmaUiPlugin, vi_mode_indicator::ViModeIndicatorPlugin};
 
 fn main() {
     // NOTE: must run before App::new() spawns any thread — they write process
@@ -68,7 +68,6 @@ fn main() {
             OrzmaWebviewPlugin {
                 orzma_assets: orzma_registry,
             },
-            ViModePlugin,
             ViModeIndicatorPlugin,
             WindowTitlePlugin,
             FocusSyncPlugin,
